Make changes to unit tests to support GRUs
The unit tests for the TV script generation project seem to expect
a tuple as the hidden state. However, in the case of GRUs, this is
not the case.

This commit changes the unit tests to deal with both a tuple as the
hidden state as well as a single value