Updated with apt variable names for validation
Throughout the exercise and solution notebooks the variable names for validation loss/losses are defined as test_loss, test_losses.
This is unnecessary confusion for beginners. Some might get confused as to what is the difference between validation and test passes when terms such as "test loss" are used for validation loss. Variable names should always be descriptive of the kind of data they represent. 
I've also updated the comment above testloader to point out that we are only temporarily using test data for the purpose of validation.