Fix ¨Visualize Sample Test Results¨ section (GPU)
Fix for ¨Visualize Sample Test Results¨ section when working with GPU:
If you are working with GPU and do not move the images back to CPU in the last part the function imshow in the last bucle will throw an error.

TypeError: can't convert CUDA tensor to numpy. Use Tensor.cpu() to copy the tensor to host memory first.

The proposed change solves the problem.